+++ CHANGELOG.md
@@ -4,3 +4,4 @@
 * Gets GitHub details from Travis & CircleCI - orta
 * Gets PR details from GitHub - orta
 * Gets Git details from local Git - orta
+* Fails when you say it's failed in  the  Dangerfile - orta